requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
setuptools
langchain-community
langchain-openai

-e .
